# [IMP] odoo.tools.sql: use generators to build SQL code