Description of the issue/feature this PR addresses:
Implementation SQL using generators to build code using less intermediary strings and just str.join() the results. Creation by overwriting __init__ instead of __new__ is also a little bit faster for creation of these small objects and reference counting will free what is needed.



---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr