@robodoo 2024-06-07
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/168485).