@kmagusiak 2024-06-07
Performance testing:
```
def create(vals):
    return SQL(",").join(SQL.identifier(x) for x in vals)

def small():
    return create("hellozor")

def long():
    return create(map(lambda i: f"i{i}", range(100000)))

def params():
    return SQL("select %s from %s where x > %s", create("x"), SQL("tab"), 5)

def params2():
    return SQL().join([SQL("select %s from %s", create("x"), SQL("tab")), SQL("where %s", 5)])

def large(level=10):
    if level == 0:
        return SQL("hello")
    if level == 1:
        return SQL("%s = %s", large(0), 2)
    if level == 10:
        return SQL("my test %s and %s%s", large(4), large(level-1), 'just a param')
    return SQL("bla " * 50 + "%s.", large(level - 1))


tests = (small, long, params, params2, large)

for test in tests:
    print(test, len(test().code))
    %timeit test()
    sql = test()
    %timeit tuple(sql)
```

Before:
<function small at 0x7f64b519f7e0> 31
5.83 μs ± 96.7 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
6.82 μs ± 147 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
<function long at 0x7f64b519f880> 888889
79.5 ms ± 2.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
66.7 ms ± 1.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
<function params at 0x7f64b519fba0> 32
1.88 μs ± 22.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
3.07 μs ± 17.5 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
<function params2 at 0x7f64b519fc40> 27
3.49 μs ± 19.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
4.79 μs ± 76.7 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
<function large at 0x7f64b543fb00> 2246
8.94 μs ± 78.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
16.8 μs ± 227 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

After:
<function small at 0x7f64b5cf37e0> 31
7.91 μs ± 742 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
3.68 μs ± 28.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
<function long at 0x7f64b519f7e0> 888889
112 ms ± 1.39 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
60.2 ms ± 1.65 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
<function params at 0x7f64b547c400> 32
3.38 μs ± 22.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.48 μs ± 23.9 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
<function params2 at 0x7f64b547c0e0> 27
4.62 μs ± 31.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.91 μs ± 7.83 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
<function large at 0x7f64b547c180> 2246
11.6 μs ± 85.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
9.02 μs ± 188 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

Creation is slower but generation of SQL is faster. In total, most of the time it's faster.